Superhirn PATCH by Florian Zeller: Using correct types in simple_math2
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10076 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/TRANSFORMATIONS/FEATUREFINDER/SUPERHIRN/simple_math2.C
@@ -101,7 +101,7 @@ void simple_math2::ITERATIVE_OUTLIER_DETECTION_BY_DIXON(vector<double>* IN){
   
   ////////////////
   // start here teh outlier detection iteratively:
-  unsigned int size;
+  size_t size;
   do{
     size = TMP.size();
     // make a outlier detection:
@@ -124,7 +124,7 @@ void simple_math2::ITERATIVE_OUTLIER_DETECTION_BY_DIXON(vector< pair<double, voi
   
   ////////////////
   // start here teh outlier detection iteratively:
-  unsigned int size;
+  size_t size;
   do{
     size = IN->size();
     // make a outlier detection:
